pyproject.toml
[project]
name = "nomad-travel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "streamlit>=1.52.2",
]
